Control modin partitions with set_npartitions as well
